+++ packages/desktop/components/popups/TestDeepLinkFormPopup.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
-    import { handleDeepLink } from '@auxiliary/deep-link'
-
+    import { handleDeepLink } from '../../lib/auxiliary/deep-link'
     import { localize } from '@core/i18n'
     import { closePopup } from '@auxiliary/popup'
     import { Button, Text, TextInput, FontWeight, TextType } from 'shared/components'
