+++ packages/desktop/lib/auxiliary/deep-link/actions/handleDeepLink.ts
