+++ packages/desktop/lib/auxiliary/deep-link/actions/initialiseDeepLinkManager.ts
