+++ packages/desktop/lib/auxiliary/deep-link/classes/deep-link-manager.class.ts
