+++ packages/desktop/lib/auxiliary/deep-link/classes/index.ts
