+++ packages/desktop/lib/auxiliary/deep-link/constants/index.ts
