+++ packages/desktop/lib/auxiliary/deep-link/constants/url-cleanup-regex.constant.ts
