+++ packages/desktop/lib/auxiliary/deep-link/enums/add-proposal-parameter.enum.ts
