+++ packages/desktop/lib/auxiliary/deep-link/enums/governance-operation.enum.ts
