+++ packages/desktop/lib/auxiliary/deep-link/enums/index.ts
