+++ packages/desktop/lib/auxiliary/deep-link/enums/send-operation-parameter.enum.ts
