+++ packages/desktop/lib/auxiliary/deep-link/enums/wallet-operation.enum.ts
