+++ packages/desktop/lib/auxiliary/deep-link/errors/amount-not-an-integer.error.ts
