+++ packages/desktop/lib/auxiliary/deep-link/errors/index.ts
@@ -1,5 +1,4 @@
 export * from './amount-not-an-integer.error'
-export * from './invalid-address.error'
 export * from './metadata-length.error'
 export * from './no-address-specified.error'
 export * from './tag-length.error'
