+++ packages/desktop/lib/auxiliary/deep-link/errors/metadata-length.error.ts
