+++ packages/desktop/lib/auxiliary/deep-link/errors/no-address-specified.error.ts
