+++ packages/desktop/lib/auxiliary/deep-link/errors/surplus-not-a-number.error.ts
