+++ packages/desktop/lib/auxiliary/deep-link/errors/unknown-asset.error.ts
