+++ packages/desktop/lib/auxiliary/deep-link/helpers/getRawAmountFromSearchParam.ts
