+++ packages/desktop/lib/auxiliary/deep-link/helpers/index.ts
