+++ packages/desktop/lib/auxiliary/deep-link/index.ts
