+++ packages/desktop/lib/auxiliary/deep-link/utils/index.ts
