+++ packages/shared/lib/core/utils/crypto/errors/index.ts
@@ -1,0 +1,1 @@
+export * from './invalid-address.error'
