+++ packages/shared/lib/core/utils/crypto/index.ts
@@ -1,3 +1,4 @@
 export * from './classes'
 export * from './constants'
+export * from './errors'
 export * from './utils'
