+++ packages/shared/lib/core/wallet/errors/missing-transaction-id.error.ts
@@ -1,4 +1,5 @@
-import { BaseError, DEFAULT_APP_ERROR_PARAMETERS } from '@core/error'
+import { BaseError } from '@core/error/classes'
+import { DEFAULT_APP_ERROR_PARAMETERS } from '@core/error/constants'
 
 export class MissingTransactionIdError extends BaseError {
     constructor() {
